feature：支持浮点数运算

+++ src/headers/lisp.h
@@ -42,7 +42,7 @@ typedef struct {
     int type;
     union {
         int err;
-        long num;
+        double num;
     };
 } lval;
